[kimchi][proof] move proof structs to a proof module
