+++ kimchi/src/lib.rs
@@ -9,6 +9,7 @@ pub mod circuits;
 pub mod error;
 pub mod linearization;
 pub mod plonk_sponge;
+pub mod proof;
 pub mod prover;
 pub mod prover_index;
 pub mod verifier;
